Fixe Uri

+++ lib/api/modle/current_temp_model.dart
@@ -1,9 +1,8 @@
-
 class CurrentWeatherModel {
    String location ;
    double temperature ;
-   double clouds ;
-   double humidity ;
+   int clouds ;
+   int humidity ;
    double wind ;
